forgedfate/requirements.txt
# Core dependencies
elasticsearch>=8.0.0,<9.0.0
aiofiles>=0.8.0
asyncio-mqtt>=0.11.0
pydantic>=1.10.0
click>=8.0.0
PyYAML>=6.0
python-dateutil>=2.8.0
requests>=2.28.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Logging and monitoring
structlog>=22.0.0
prometheus-client>=0.14.0

# Configuration management
python-dotenv>=0.19.0
configparser>=5.3.0

# Networking and protocols
websockets>=10.0
scapy>=2.4.5

# Development dependencies (optional)
pytest>=6.0.0
pytest-asyncio>=0.18.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
pre-commit>=2.15.0

# Documentation (optional)
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.17.0
